update doc viewer template and style edit button
