+++ apps/daffio/src/app/docs/components/doc-viewer/doc-viewer.component.scss
@@ -31,15 +31,44 @@
 		}
 	}
 
-	&__table-of-contents {
+	&__aside-wrapper {
 		display: none;
 
 		@include daff.breakpoint(small-laptop) {
+			align-self: flex-start;
 			display: block;
+			position: sticky;
+			top: calc(var(--daff-sidebar-side-fixed-top-shift) + 3rem);
 			flex-grow: 1;
 			max-width: 15rem;
 			width: 100%;
+			height: calc(100vh - var(--daff-sidebar-side-fixed-top-shift) - 6rem);
 		}
+	}
+
+	&__aside-title {
+		font-weight: 500;
+		margin: 0 0 1rem;
+	}
+
+	&__toc {
+		overflow-y: auto;
+		max-height: 65vh;
+	}
+
+	&__toc-links {
+		display: flex;
+		flex-direction: column;
+		gap: 1rem;
+		margin: 1rem 0 0;
+		padding: 1rem 0 0;
+	}
+
+	&__edit-button {
+		display: flex;
+		align-items: center;
+		gap: 0.5rem;
+		text-decoration: none;
 	}
 
 	&__sidebar-activators {
