+++ apps/daffio/src/app/docs/components/scroll-to-top/scroll-to-top.component.scss
@@ -9,6 +9,8 @@
 		appearance: none;
 		background: none;
 		border: none;
+		margin: 0;
+		padding: 0;
 		position: relative;
 	}
 }
