--- apps/daffio/src/app/docs/components/table-of-contents/link/link.component.html
@@ -1,11 +1,0 @@
-<ng-container *ngrxLet="activeHeaderService.fragment$; let fragment">
-	@for (entry of tableOfContents; track entry.slug) {
-		<a
-			class="daffio-docs-table-of-contents-link level-{{entry.lvl}}"
-			routerLink="./"
-			queryParamsHandling="merge"
-			[fragment]="entry.slug"
-			[class.active]="fragment === entry.slug"
-		>{{ entry.content }}</a>
-	}
-</ng-container>
