--- apps/daffio/src/app/docs/components/table-of-contents/link/link.component.scss
@@ -1,52 +1,0 @@
-@use 'typography' as t;
-
-// stylelint-disable selector-class-pattern
-:host {
-	display: block;
-	max-height: calc(100vh - var(--daff-sidebar-side-fixed-top-shift) - 9rem);
-	overflow-y: auto;
-
-	&.in-sidebar {
-		.daffio-docs-table-of-contents-link {
-			&::before {
-				left: -1rem;
-			}
-		}
-	}
-}
-
-.daffio-docs-table-of-contents-link {
-	@include t.text-truncate();
-	display: block;
-	font-size: 1rem;
-	line-height: 1.25rem;
-	text-decoration: none;
-	position: relative;
-	padding: 0.5rem 0 0.5rem 0.75rem;
-
-	&::before {
-		content: '';
-		position: absolute;
-		top: 0.5rem;
-		bottom: 0;
-		left: 0;
-		width: 0.125rem;
-		height: 1.25rem;
-	}
-
-	&.active {
-		font-weight: 500;
-	}
-
-	&.level-1 {
-		display: none;
-	}
-
-	&.level-3 {
-		padding-left: 1.5rem;
-	}
-
-	&.level-4 {
-		padding-left: 2.5rem;
-	}
-}
