+++ apps/daffio/src/app/docs/components/table-of-contents/table-of-contents.component.scss
@@ -1,22 +1,50 @@
-@use 'utilities' as daff;
+@use 'typography' as t;
 
+// stylelint-disable selector-class-pattern
 :host {
 	display: block;
+
+	&.in-sidebar {
+		.daffio-docs-table-of-contents-link {
+			&::before {
+				left: -1rem;
+			}
+		}
+	}
 }
 
-.daffio-docs-table-of-contents {
-	position: sticky;
-	top: 6rem;
-	max-width: 15rem;
-	width: 100%;
+.daffio-docs-table-of-contents-link {
+	@include t.text-truncate();
+	display: block;
+	font-size: 1rem;
+	line-height: 1.25rem;
+	text-decoration: none;
+	position: relative;
+	padding: 0.5rem 0 0.5rem 0.75rem;
 
-	&__title {
-		font-weight: 500;
-		margin: 0 0 1.5rem;
+	&::before {
+		content: '';
+		position: absolute;
+		top: 0.5rem;
+		bottom: 0;
+		left: 0;
+		width: 0.125rem;
+		height: 1.25rem;
 	}
 
-	&__scroll-to-top {
-		display: block;
-		margin: 1rem 0 0;
+	&.active {
+		font-weight: 500;
+	}
+
+	&.level-1 {
+		display: none;
+	}
+
+	&.level-3 {
+		padding-left: 1.5rem;
+	}
+
+	&.level-4 {
+		padding-left: 2.5rem;
 	}
 }
