+++ apps/daffio/src/app/docs/containers/toc-sidebar-content/toc-sidebar-content.component.html
@@ -1,1 +1,1 @@
-<daffio-docs-table-of-contents-link [tableOfContents]="toc()" [inSidebar]="true"></daffio-docs-table-of-contents-link>
+<daffio-docs-table-of-contents [tableOfContents]="toc()" [inSidebar]="true"></daffio-docs-table-of-contents>
