+++ apps/daffio/src/scss/component-themes.scss
@@ -10,7 +10,7 @@
 @use '../app/core/sidebar/components/docs/footer/footer-theme' as docs-sidebar-footer;
 @use '../app/core/header/components/header/header-theme' as header;
 @use '../app/docs/components/scroll-to-top/scroll-to-top-theme' as scroll-to-top;
-@use '../app/docs/components/table-of-contents/link/link-theme' as docs-toc-link;
+@use '../app/docs/components/table-of-contents/table-of-contents-theme' as docs-toc;
 @use '../app/docs/components/member-heading/member-heading-theme' as member-heading;
 @use '../app/docs/api/components/fragments/methods/methods-theme.scss' as api-methods-fragment;
 @use '../app/docs/api/components/fragments/params/params-theme.scss' as api-params-fragment;
@@ -19,6 +19,7 @@
 @use '../app/docs/api/components/method-block/method-block-theme' as api-method-block;
 @use '../app/docs/api/components/api-item-label/api-item-label-theme' as api-item-label;
 @use '../app/docs/components/heading-link/heading-link-theme' as docs-heading-link;
+@use '../app/docs/components/doc-viewer/doc-viewer-theme' as doc-viewer;
 
 @mixin component-themes($theme) {
 	@include home-hero.daffio-home-hero-theme($theme);
@@ -35,7 +36,7 @@
 	@include marketing-sidebar-footer.daffio-marketing-sidebar-footer-theme($theme);
 	@include docs-sidebar-footer.daffio-docs-sidebar-footer-theme($theme);
 	@include scroll-to-top.daffio-docs-scroll-to-top-theme($theme);
-	@include docs-toc-link.daffio-docs-table-of-contents-link-theme($theme);
+	@include docs-toc.daffio-docs-table-of-contents-theme($theme);
 	@include api-methods-fragment.daffio-api-methods-fragment-theme($theme);
 	@include api-params-fragment.daffio-api-params-fragment-theme($theme);
 	@include api-method-block.daffio-method-block-theme($theme);
@@ -43,4 +44,5 @@
 	@include api-outputs-fragment.daffio-api-outputs-fragment-theme($theme);
 	@include api-item-label.daffio-api-item-label-theme($theme);
 	@include docs-heading-link.daffio-docs-heading-link-theme($theme);
+	@include doc-viewer.daffio-doc-viewer-theme($theme);
 }
